Complete GUI restoration.
All panels resized to 800x480.
Added functionality of pnlTakePill and pnlInsertPill with the right
position of buttons. 
For an example application, it is possible to use small button at the
bottom of the window. Here the logic:
0) Alarm Window
1) pnlTakePill--> "avanti" is a return home button
2) pnlInsertPill--> ok doesn't work fine

Actually it works but need more structured code to resize images and
text considering a generic prescription.
A link to Excel database is also needed.

pnlInsertPill still not working fine with "switch logic".